requirements.txt
pillow
django
django-rest-framework
django-cors-headers
mysqlclient